minor changes, and made it compile on mac os x

+++ trunk/include/mm.h
@@ -46,8 +46,8 @@ void printk (const char *, ...);
 #ifdef __cplusplus
 } extern "C++" {
 
-void* operator new(unsigned int size);
-void* operator new[](unsigned int size);
+void* operator new(size_t size);
+void* operator new[](size_t size);
 void operator delete(void * ptr);
 void operator delete[](void * ptr);
